Split out formulae detect step.
While we're here, also cleanup some output formatting issues and dead
code I noticed along the way.
